# Conflicts:
#	fiit-mobile/app/src/main/java/ar/edu/itba/hci/fiit_mobile/Components/RoutineCard.kt
#	fiit-mobile/app/src/main/java/ar/edu/itba/hci/fiit_mobile/ui/views/FavsView.kt

+++ fiit-mobile/app/src/main/java/ar/edu/itba/hci/fiit_mobile/data/network/model/user/NetworkUser.kt
@@ -9,7 +9,7 @@ data class NetworkUser (
     @SerializedName("username"     ) var username     : String,
     @SerializedName("gender"       ) var gender       : String? = null,
     @SerializedName("avatarUrl"    ) var avatarUrl    : String? = null,
-    @SerializedName("date"         ) var date         : Int? = null,
+    @SerializedName("date"         ) var date         : Long? = null,
     @SerializedName("lastActivity" ) var lastActivity : Int?    = null
 
 )
