+++ fiit-mobile/app/src/main/java/ar/edu/itba/hci/fiit_mobile/data/network/model/user/NetworkUserContent.kt
@@ -15,7 +15,7 @@ data class NetworkUserContent (
     @SerializedName("phone"        ) var phone        : String?  = null,
     @SerializedName("avatarUrl"    ) var avatarUrl    : String?  = null,
     @SerializedName("metadata"     ) var metadata     : String?  = null,
-    @SerializedName("date"         ) var date         : Int?     = null,
+    @SerializedName("date"         ) var date         : Long?     = null,
     @SerializedName("lastActivity" ) var lastActivity : Int?     = null,
     @SerializedName("verified"     ) var verified     : Boolean? = null
